requirements.txt
fastapi[standard]
uvicorn
Flask==2.3.3
Flask-pymongo==2.3.0
gunicorn==21.2.0
python-jose[cryptography]
python-dotenv
dnspython
mongoengine
requests